refact: moved Role URIs to VitroVocabulary
